+++ CoreLib/IndexedDataAccess.h
@@ -20,8 +20,8 @@ public:
 	void				FreeIterator(CIndexTreeIterator* pcIter);
 
 protected:
-	BOOL	Put(void* pvKey, int iKeySize, void* pvData, int iDataSize);
-	BOOL	Get(void* pvKey, int iKeySize, void* pvData, int* piDataSize);
+	BOOL	Put(void* pvKey, int iKeySize, void* pvData, size_t iDataSize);
+	BOOL	Get(void* pvKey, int iKeySize, void* pvData, size_t* piDataSize);
 	int		DataSize(void* pvKey, int iKeySize);
 	BOOL	Remove(void* pvKey, int iKeySize);
 	BOOL	Has(void* pvKey, int iKeySize);
